Decompose verifiable_credential_1_1.rs (#321)
