+++ crates/web5/src/credentials/jwt_payload_vc.rs
@@ -1,0 +1,32 @@
+use std::time::SystemTime;
+
+use serde::{Deserialize, Serialize};
+
+use super::{credential_subject::CredentialSubject, issuer::Issuer};
+use crate::rfc3339::{deserialize_optional_system_time, serialize_optional_system_time};
+
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct JwtPayloadVerifiableCredential {
+    #[serde(rename = "@context")]
+    pub context: Vec<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub id: Option<String>,
+    #[serde(rename = "type")]
+    pub r#type: Vec<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub issuer: Option<Issuer>,
+    #[serde(
+        rename = "issuanceDate",
+        serialize_with = "serialize_optional_system_time",
+        deserialize_with = "deserialize_optional_system_time"
+    )]
+    pub issuance_date: Option<SystemTime>,
+    #[serde(
+        rename = "expirationDate",
+        serialize_with = "serialize_optional_system_time",
+        deserialize_with = "deserialize_optional_system_time"
+    )]
+    pub expiration_date: Option<SystemTime>,
+    #[serde(skip_serializing_if = "Option::is_none", rename = "credentialSubject")]
+    pub credential_subject: Option<CredentialSubject>,
+}
