+++ crates/web5/src/credentials/verifiable_credential_1_1.rs
@@ -1,105 +1,21 @@
-use super::CredentialError;
-use crate::dids::data_model::document::FindVerificationMethodOptions;
-use crate::errors::{Result, Web5Error};
-use crate::json::{FromJson, JsonObject, ToJson};
+use super::data_model_validation::validate_vc_data_model;
+use super::decode::decode;
+use super::CredentialSubject;
+use super::Issuer;
+
+use crate::dids::bearer_did::BearerDid;
+use crate::errors::Result;
+use crate::json::{FromJson, ToJson};
 use crate::rfc3339::{
     deserialize_optional_system_time, deserialize_system_time, serialize_optional_system_time,
     serialize_system_time,
 };
-use crate::{
-    crypto::dsa::{ed25519::Ed25519Verifier, Signer, Verifier},
-    dids::{
-        bearer_did::BearerDid,
-        did::Did,
-        resolution::{
-            resolution_metadata::ResolutionMetadataError, resolution_result::ResolutionResult,
-        },
-    },
-};
-use core::fmt;
-use josekit::{
-    jws::{
-        alg::eddsa::EddsaJwsAlgorithm as JosekitEddsaJwsAlgorithm,
-        JwsAlgorithm as JosekitJwsAlgorithm, JwsHeader, JwsSigner as JosekitJwsSigner,
-        JwsVerifier as JosekitJwsVerifier,
-    },
-    jwt::JwtPayload,
-    JoseError as JosekitError,
-};
+
 use serde::{Deserialize, Serialize};
-use std::{
-    fmt::{Display, Formatter},
-    sync::Arc,
-    time::SystemTime,
-};
-use uuid::Uuid;
+use std::time::SystemTime;
 
 pub const BASE_CONTEXT: &str = "https://www.w3.org/2018/credentials/v1";
 pub const BASE_TYPE: &str = "VerifiableCredential";
-
-#[derive(Serialize, Deserialize, Debug, Default, Clone, PartialEq)]
-pub struct ObjectIssuer {
-    pub id: String,
-    pub name: String,
-    #[serde(flatten)]
-    pub additional_properties: Option<JsonObject>,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-#[serde(untagged)]
-pub enum Issuer {
-    String(String),
-    Object(ObjectIssuer),
-}
-
-impl FromJson for Issuer {}
-impl ToJson for Issuer {}
-
-impl<I> From<I> for Issuer
-where
-    I: Into<String>,
-{
-    fn from(s: I) -> Self {
-        Issuer::String(s.into())
-    }
-}
-
-impl Display for Issuer {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        match self {
-            Issuer::String(s) => write!(f, "{}", s),
-            Issuer::Object(ni) => write!(f, "{}", ni.id),
-        }
-    }
-}
-
-#[derive(Serialize, Deserialize, Debug, Default, Clone, PartialEq)]
-pub struct CredentialSubject {
-    pub id: String,
-    #[serde(flatten)]
-    pub additional_properties: Option<JsonObject>,
-}
-
-impl FromJson for CredentialSubject {}
-impl ToJson for CredentialSubject {}
-
-impl<I> From<I> for CredentialSubject
-where
-    I: Into<String>,
-{
-    fn from(s: I) -> Self {
-        CredentialSubject {
-            id: s.into(),
-            ..Default::default()
-        }
-    }
-}
-
-impl Display for CredentialSubject {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        write!(f, "{}", self.id)
-    }
-}
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct VerifiableCredential {
@@ -143,212 +59,12 @@ impl VerifiableCredential {
         credential_subject: CredentialSubject,
         options: Option<VerifiableCredentialCreateOptions>,
     ) -> Result<Self> {
-        if issuer.to_string().is_empty() {
-            return Err(Web5Error::Parameter(String::from(
-                "issuer id must not be empty",
-            )));
-        }
-
-        if let Issuer::Object(ref named_issuer) = issuer {
-            if named_issuer.name.is_empty() {
-                return Err(Web5Error::Parameter(String::from(
-                    "named issuer name must not be empty",
-                )));
-            }
-        }
-
-        if credential_subject.to_string().is_empty() {
-            return Err(Web5Error::Parameter(String::from(
-                "subject id must not be empty",
-            )));
-        }
-
-        if Did::parse(&issuer.to_string()).is_err() {
-            return Err(Web5Error::Parameter(
-                "issuer must be a valid DID URI".to_string(),
-            ));
-        }
-
-        if Did::parse(&credential_subject.to_string()).is_err() {
-            return Err(Web5Error::Parameter(
-                "credential subject must be a valid DID URI".to_string(),
-            ));
-        }
-
-        let options = options.unwrap_or_default();
-
-        let context = {
-            let mut contexts = options
-                .context
-                .unwrap_or_else(|| vec![BASE_CONTEXT.to_string()]);
-
-            if !contexts.contains(&BASE_CONTEXT.to_string()) {
-                contexts.insert(0, BASE_CONTEXT.to_string());
-            }
-
-            contexts
-        };
-
-        let r#type = {
-            let mut types = options
-                .r#type
-                .unwrap_or_else(|| vec![BASE_TYPE.to_string()]);
-
-            if !types.contains(&BASE_TYPE.to_string()) {
-                types.insert(0, BASE_TYPE.to_string());
-            }
-
-            types
-        };
-
-        let id = options
-            .id
-            .unwrap_or_else(|| format!("urn:uuid:{}", Uuid::new_v4()));
-
-        Ok(Self {
-            context,
-            id,
-            r#type,
-            issuer,
-            issuance_date: options.issuance_date.unwrap_or_else(SystemTime::now),
-            expiration_date: options.expiration_date,
-            credential_subject,
-        })
+        super::create::create_vc(issuer, credential_subject, options)
     }
 
     // this function currently only supports Ed25519
     pub fn from_vc_jwt(vc_jwt: &str, verify: bool) -> Result<Self> {
-        let header = josekit::jwt::decode_header(vc_jwt)
-            .map_err(|_| Web5Error::Parameter("failed to decode vc-jwt jose header".to_string()))?;
-
-        let kid = header
-            .claim("kid")
-            .and_then(serde_json::Value::as_str)
-            .ok_or(CredentialError::MissingKid)?;
-
-        if kid.is_empty() {
-            return Err(CredentialError::MissingKid.into());
-        }
-
-        let jwt_payload = if verify {
-            let did = Did::parse(kid)?;
-
-            let resolution_result = ResolutionResult::resolve(&did.uri);
-            if let Some(err) = resolution_result.resolution_metadata.error.clone() {
-                return Err(err.into());
-            }
-
-            let public_key_jwk = resolution_result
-                .document
-                .ok_or(ResolutionMetadataError::InternalError)?
-                .find_verification_method(FindVerificationMethodOptions {
-                    verification_method_id: Some(kid.to_string()),
-                })?
-                .public_key_jwk;
-
-            let jose_verifier = &JoseVerifier {
-                kid: kid.to_string(),
-                verifier: Arc::new(Ed25519Verifier::new(public_key_jwk)),
-            };
-
-            let (jwt_payload, _) = josekit::jwt::decode_with_verifier(vc_jwt, jose_verifier)
-                .map_err(|e| {
-                    Web5Error::Crypto(format!("vc-jwt failed cryptographic verification {}", e))
-                })?;
-
-            jwt_payload
-        } else {
-            let (jwt_payload, _) = josekit::jwt::decode_with_verifier(
-                vc_jwt,
-                &JoseVerifierAlwaysTrue {
-                    kid: kid.to_string(),
-                },
-            )
-            .map_err(|e| Web5Error::Crypto(format!("vc-jwt failed to decode payload {}", e)))?;
-
-            jwt_payload
-        };
-
-        let vc_claim = jwt_payload
-            .claim("vc")
-            .ok_or(CredentialError::MissingClaim("vc".to_string()))?;
-        let vc_payload =
-            serde_json::from_value::<JwtPayloadVerifiableCredential>(vc_claim.clone())?;
-
-        // registered claims checks
-        let jti = jwt_payload
-            .jwt_id()
-            .ok_or(CredentialError::MissingClaim("jti".to_string()))?;
-        let iss = jwt_payload
-            .issuer()
-            .ok_or(CredentialError::MissingClaim("issuer".to_string()))?;
-        let sub = jwt_payload
-            .subject()
-            .ok_or(CredentialError::MissingClaim("subject".to_string()))?;
-        let nbf = jwt_payload
-            .not_before()
-            .ok_or(CredentialError::MissingClaim("not_before".to_string()))?;
-        let exp = jwt_payload.expires_at();
-
-        if let Some(id) = &vc_payload.id {
-            if id != jti {
-                return Err(CredentialError::ClaimMismatch("id".to_string()).into());
-            }
-        }
-
-        if let Some(issuer) = &vc_payload.issuer {
-            let vc_issuer = issuer.to_string();
-            if iss != vc_issuer {
-                return Err(CredentialError::ClaimMismatch("issuer".to_string()).into());
-            }
-        }
-
-        if let Some(credential_subject) = &vc_payload.credential_subject {
-            if sub != credential_subject.id {
-                return Err(CredentialError::ClaimMismatch("subject".to_string()).into());
-            }
-        }
-
-        if let Some(vc_payload_expiration_date) = vc_payload.expiration_date {
-            match exp {
-                None => {
-                    return Err(CredentialError::MisconfiguredExpirationDate(
-                        "VC has expiration date but no exp in registered claims".to_string(),
-                    )
-                    .into());
-                }
-                Some(exp) => {
-                    let difference = vc_payload_expiration_date
-                        .duration_since(exp)
-                        .unwrap_or_else(|_| {
-                            exp.duration_since(vc_payload_expiration_date).unwrap()
-                        });
-
-                    if difference.as_secs() > 0 {
-                        return Err(
-                            CredentialError::ClaimMismatch("expiration_date".to_string()).into(),
-                        );
-                    }
-                }
-            }
-        }
-
-        let vc_issuer = vc_payload.issuer.unwrap_or(Issuer::String(iss.to_string()));
-
-        let vc_credential_subject = vc_payload.credential_subject.unwrap_or(CredentialSubject {
-            id: sub.to_string(),
-            additional_properties: None,
-        });
-
-        let vc = VerifiableCredential {
-            context: vc_payload.context,
-            id: jti.to_string(),
-            r#type: vc_payload.r#type,
-            issuer: vc_issuer,
-            issuance_date: nbf,
-            expiration_date: exp,
-            credential_subject: vc_credential_subject,
-        };
+        let vc = decode(vc_jwt, verify)?;
 
         if verify {
             validate_vc_data_model(&vc)?;
@@ -357,638 +73,43 @@ impl VerifiableCredential {
         Ok(vc)
     }
 
-    fn sign_with_signer(&self, key_id: &str, signer: Arc<dyn Signer>) -> Result<String> {
-        let mut payload = JwtPayload::new();
-        let vc_claim = JwtPayloadVerifiableCredential {
-            context: self.context.clone(),
-            id: Some(self.id.clone()),
-            r#type: self.r#type.clone(),
-            issuer: Some(self.issuer.clone()),
-            issuance_date: Some(self.issuance_date),
-            expiration_date: self.expiration_date,
-            credential_subject: Some(self.credential_subject.clone()),
-        };
-        payload
-            .set_claim("vc", Some(serde_json::to_value(vc_claim)?))
-            .map_err(|e| Web5Error::Unknown(format!("failed to set claim {}", e)))?;
-        payload.set_issuer(self.issuer.to_string());
-        payload.set_jwt_id(&self.id);
-        payload.set_subject(&self.credential_subject.id);
-        payload.set_not_before(&self.issuance_date);
-        payload.set_issued_at(&SystemTime::now());
-        if let Some(exp) = &self.expiration_date {
-            payload.set_expires_at(exp)
-        }
-
-        let jose_signer = JoseSigner {
-            kid: key_id.to_string(),
-            signer,
-        };
-
-        let mut header = JwsHeader::new();
-        header.set_token_type("JWT");
-        let vc_jwt = josekit::jwt::encode_with_signer(&payload, &header, &jose_signer)
-            .map_err(|e| Web5Error::Crypto(format!("failed to sign vc-jwt {}", e)))?;
-
-        Ok(vc_jwt)
-    }
-
     pub fn sign(
         &self,
         bearer_did: &BearerDid,
         verification_method_id: Option<String>,
     ) -> Result<String> {
-        // contains() because the issuer may also include a DID Fragment
-        if self.issuer.to_string() != bearer_did.did.uri.clone() {
-            return Err(Web5Error::Parameter(format!(
-                "bearer_did uri {} does not match issuer {}",
-                bearer_did.did.uri.clone(),
-                self.issuer
-            )));
-        }
-
-        let verification_method_id = verification_method_id
-            .unwrap_or_else(|| bearer_did.document.verification_method[0].id.clone());
-
-        let is_assertion_method =
-            if let Some(assertion_methods) = &bearer_did.document.assertion_method {
-                assertion_methods.contains(&verification_method_id)
-            } else {
-                false
-            };
-
-        if !is_assertion_method {
-            return Err(Web5Error::Parameter(format!(
-                "verification_method_id {} is not an assertion_method",
-                verification_method_id
-            )));
-        }
-
-        let signer = bearer_did.get_signer(&verification_method_id)?;
-        self.sign_with_signer(&verification_method_id, signer)
-    }
-}
-
-fn validate_vc_data_model(vc: &VerifiableCredential) -> Result<()> {
-    // Required fields ["@context", "id", "type", "issuer", "issuanceDate", "credentialSubject"]
-    if vc.id.is_empty() {
-        return Err(CredentialError::DataModelValidationError("missing id".to_string()).into());
-    }
-
-    if vc.context.is_empty() || vc.context[0] != BASE_CONTEXT {
-        return Err(
-            CredentialError::DataModelValidationError("missing context".to_string()).into(),
-        );
-    }
-
-    if vc.r#type.is_empty() || vc.r#type[0] != BASE_TYPE {
-        return Err(CredentialError::DataModelValidationError("missing type".to_string()).into());
-    }
-
-    if vc.issuer.to_string().is_empty() {
-        return Err(CredentialError::DataModelValidationError("missing issuer".to_string()).into());
-    }
-
-    if vc.credential_subject.id.is_empty() {
-        return Err(CredentialError::DataModelValidationError(
-            "missing credential subject".to_string(),
-        )
-        .into());
-    }
-
-    let now = SystemTime::now();
-    if vc.issuance_date > now {
-        return Err(CredentialError::DataModelValidationError(
-            "issuance date in future".to_string(),
-        )
-        .into());
-    }
-
-    // Validate expiration date if it exists
-    if let Some(expiration_date) = &vc.expiration_date {
-        if expiration_date < &now {
-            return Err(CredentialError::DataModelValidationError(
-                "credential expired".to_string(),
-            )
-            .into());
-        }
-    }
-
-    // TODO: Add validations to credential_status, credential_schema, and evidence once they are added to the VcDataModel
-    // https://github.com/TBD54566975/web5-rs/issues/112
-
-    Ok(())
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone)]
-struct JwtPayloadVerifiableCredential {
-    #[serde(rename = "@context")]
-    context: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    id: Option<String>,
-    #[serde(rename = "type")]
-    r#type: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    issuer: Option<Issuer>,
-    #[serde(
-        rename = "issuanceDate",
-        serialize_with = "serialize_optional_system_time",
-        deserialize_with = "deserialize_optional_system_time"
-    )]
-    issuance_date: Option<SystemTime>,
-    #[serde(
-        rename = "expirationDate",
-        serialize_with = "serialize_optional_system_time",
-        deserialize_with = "deserialize_optional_system_time"
-    )]
-    expiration_date: Option<SystemTime>,
-    #[serde(skip_serializing_if = "Option::is_none", rename = "credentialSubject")]
-    credential_subject: Option<CredentialSubject>,
-}
-
-#[derive(Clone)]
-struct JoseSigner {
-    pub kid: String,
-    pub signer: Arc<dyn Signer>,
-}
-
-impl JosekitJwsSigner for JoseSigner {
-    fn algorithm(&self) -> &dyn JosekitJwsAlgorithm {
-        &JosekitEddsaJwsAlgorithm::Eddsa
-    }
-
-    fn key_id(&self) -> Option<&str> {
-        Some(&self.kid)
-    }
-
-    fn signature_len(&self) -> usize {
-        64
-    }
-
-    fn sign(&self, message: &[u8]) -> core::result::Result<Vec<u8>, JosekitError> {
-        self.signer
-            .sign(message)
-            // 🚧 improve error message semantics
-            .map_err(|err| JosekitError::InvalidSignature(err.into()))
-    }
-
-    fn box_clone(&self) -> Box<dyn JosekitJwsSigner> {
-        Box::new(self.clone())
-    }
-}
-
-impl core::fmt::Debug for JoseSigner {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Signer").field("kid", &self.kid).finish()
-    }
-}
-
-#[derive(Clone)]
-struct JoseVerifier {
-    pub kid: String,
-    pub verifier: Arc<dyn Verifier>,
-}
-
-impl JosekitJwsVerifier for JoseVerifier {
-    fn algorithm(&self) -> &dyn JosekitJwsAlgorithm {
-        &JosekitEddsaJwsAlgorithm::Eddsa
-    }
-
-    fn key_id(&self) -> Option<&str> {
-        Some(self.kid.as_str())
-    }
-
-    fn verify(&self, message: &[u8], signature: &[u8]) -> core::result::Result<(), JosekitError> {
-        self.verifier
-            .verify(message, signature)
-            .map_err(|e| JosekitError::InvalidSignature(e.into()))
-    }
-
-    fn box_clone(&self) -> Box<dyn JosekitJwsVerifier> {
-        Box::new(self.clone())
-    }
-}
-
-impl core::fmt::Debug for JoseVerifier {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Verifier").field("kid", &self.kid).finish()
-    }
-}
-
-#[derive(Clone)]
-struct JoseVerifierAlwaysTrue {
-    pub kid: String,
-}
-impl JosekitJwsVerifier for JoseVerifierAlwaysTrue {
-    fn algorithm(&self) -> &dyn JosekitJwsAlgorithm {
-        &JosekitEddsaJwsAlgorithm::Eddsa
-    }
-
-    fn key_id(&self) -> Option<&str> {
-        Some(self.kid.as_str())
-    }
-
-    fn verify(&self, _message: &[u8], _signature: &[u8]) -> core::result::Result<(), JosekitError> {
-        Ok(())
-    }
-
-    fn box_clone(&self) -> Box<dyn JosekitJwsVerifier> {
-        Box::new(self.clone())
-    }
-}
-impl core::fmt::Debug for JoseVerifierAlwaysTrue {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Verifier").field("kid", &"").finish()
+        super::sign::sign_with_did(self, bearer_did, verification_method_id)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::json::JsonValue;
-    use crate::{test_helpers::UnitTestSuite, test_name};
-    use lazy_static::lazy_static;
-    use regex::Regex;
-    use std::collections::HashMap;
-
-    const ISSUER_DID_URI: &str = "did:web:tbd.website";
-    const SUBJECT_DID_URI: &str = "did:dht:qgmmpyjw5hwnqfgzn7wmrm33ady8gb8z9ideib6m9gj4ys6wny8y";
-
-    fn issuer() -> Issuer {
-        Issuer::from(ISSUER_DID_URI)
-    }
-    fn credential_subject() -> CredentialSubject {
-        CredentialSubject::from(SUBJECT_DID_URI)
-    }
-
-    mod create {
-        use super::*;
-
-        lazy_static! {
-            static ref TEST_SUITE: UnitTestSuite =
-                UnitTestSuite::new("verifiable_credential_1_1_create");
-        }
-
-        #[test]
-        fn z_assert_all_suite_cases_covered() {
-            // fn name prefixed with `z_*` b/c rust test harness executes in alphabetical order,
-            // unless intentionally executed with "shuffle" https://doc.rust-lang.org/rustc/tests/index.html#--shuffle
-            // this may not work if shuffled or if test list grows to the extent of 100ms being insufficient wait time
-
-            // wait 100ms to be last-in-queue of mutex lock
-            std::thread::sleep(std::time::Duration::from_millis(100));
-
-            TEST_SUITE.assert_coverage()
-        }
-
-        #[test]
-        fn test_default_context_added_if_not_supplied() {
-            TEST_SUITE.include(test_name!());
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), None).unwrap();
-
-            assert_eq!(vc.context, vec![BASE_CONTEXT]);
-        }
-
-        #[test]
-        fn test_default_context_not_duplicated_if_supplied() {
-            TEST_SUITE.include(test_name!());
-
-            let options = Some(VerifiableCredentialCreateOptions {
-                context: Some(vec![BASE_CONTEXT.to_string()]),
-                ..Default::default()
-            });
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), options).unwrap();
-
-            assert_eq!(vc.context, vec![BASE_CONTEXT]);
-        }
-
-        #[test]
-        fn test_developer_provided_context_appended_to_default() {
-            TEST_SUITE.include(test_name!());
-
-            let custom_context = "https://example.com/custom-context";
-            let options = Some(VerifiableCredentialCreateOptions {
-                context: Some(vec![custom_context.to_string()]),
-                ..Default::default()
-            });
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), options).unwrap();
-
-            assert_eq!(vc.context, vec![BASE_CONTEXT, custom_context]);
-        }
-
-        #[test]
-        fn test_default_type_added_if_not_supplied() {
-            TEST_SUITE.include(test_name!());
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), None).unwrap();
-
-            assert_eq!(vc.r#type, vec![BASE_TYPE]);
-        }
-
-        #[test]
-        fn test_default_type_not_duplicated_if_supplied() {
-            TEST_SUITE.include(test_name!());
-
-            let options = Some(VerifiableCredentialCreateOptions {
-                r#type: Some(vec![BASE_TYPE.to_string()]),
-                ..Default::default()
-            });
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), options).unwrap();
-
-            assert_eq!(vc.r#type, vec![BASE_TYPE]);
-        }
-
-        #[test]
-        fn test_developer_provided_type_appended_to_default() {
-            TEST_SUITE.include(test_name!());
-
-            let custom_type = "CustomType";
-            let options = Some(VerifiableCredentialCreateOptions {
-                r#type: Some(vec![custom_type.to_string()]),
-                ..Default::default()
-            });
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), options).unwrap();
-
-            assert_eq!(vc.r#type, vec![BASE_TYPE, custom_type]);
-        }
-
-        #[test]
-        fn test_id_generated_if_not_supplied() {
-            TEST_SUITE.include(test_name!());
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), None).unwrap();
-
-            let uuid_regex = Regex::new(r"^urn:uuid:[0-9a-fA-F-]{36}$").unwrap();
-            assert!(uuid_regex.is_match(&vc.id));
-        }
-
-        #[test]
-        fn test_id_must_be_set_if_supplied() {
-            TEST_SUITE.include(test_name!());
-
-            let custom_id = "custom-id";
-            let options = Some(VerifiableCredentialCreateOptions {
-                id: Some(custom_id.to_string()),
-                ..Default::default()
-            });
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), options).unwrap();
-
-            assert_eq!(vc.id, custom_id);
-        }
-
-        #[test]
-        fn test_issuer_string_must_not_be_empty() {
-            TEST_SUITE.include(test_name!());
-
-            let empty_issuer = Issuer::from("");
-            let result = VerifiableCredential::create(empty_issuer, credential_subject(), None);
-
-            match result {
-                Err(Web5Error::Parameter(err_msg)) => {
-                    assert_eq!(err_msg, "issuer id must not be empty");
-                }
-                _ => panic!("Expected Web5Error::Parameter with specific error message"),
-            };
-        }
-
-        #[test]
-        fn test_issuer_string_must_be_set() {
-            TEST_SUITE.include(test_name!());
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), None).unwrap();
-
-            assert_eq!(vc.issuer, issuer());
-        }
-
-        #[test]
-        fn test_issuer_string_must_be_valid_did() {
-            TEST_SUITE.include(test_name!());
-
-            let result = VerifiableCredential::create(
-                Issuer::String("did:invalid-123".to_string()),
-                credential_subject(),
-                None,
-            );
-
-            match result {
-                Err(Web5Error::Parameter(err_msg)) => {
-                    assert_eq!(err_msg, "issuer must be a valid DID URI")
-                }
-                _ => panic!("Expected Web5Error::Parameter, but got: {:?}", result),
-            };
-        }
-
-        #[test]
-        fn test_issuer_object_id_must_not_be_empty() {
-            TEST_SUITE.include(test_name!());
-
-            let issuer = Issuer::Object(ObjectIssuer {
-                id: "".to_string(),
-                name: "Example Name".to_string(),
-                additional_properties: None,
-            });
-
-            let result = VerifiableCredential::create(issuer, credential_subject(), None);
-
-            match result {
-                Err(Web5Error::Parameter(err_msg)) => {
-                    assert_eq!(err_msg, "issuer id must not be empty");
-                }
-                _ => panic!("Expected Web5Error::Parameter with specific error message"),
-            };
-        }
-
-        #[test]
-        fn test_issuer_object_id_must_be_valid_did() {
-            TEST_SUITE.include(test_name!());
-
-            let result = VerifiableCredential::create(
-                issuer(),
-                CredentialSubject {
-                    id: "did:something-invalid".to_string(),
-                    ..Default::default()
-                },
-                None,
-            );
-
-            match result {
-                Err(Web5Error::Parameter(err_msg)) => {
-                    assert_eq!(err_msg, "credential subject must be a valid DID URI")
-                }
-                _ => panic!("Expected Web5Error::Parameter, but got: {:?}", result),
-            };
-        }
-
-        #[test]
-        fn test_issuer_object_name_must_not_be_empty() {
-            TEST_SUITE.include(test_name!());
-
-            let issuer = Issuer::Object(ObjectIssuer {
-                id: ISSUER_DID_URI.to_string(),
-                name: "".to_string(),
-                additional_properties: None,
-            });
-
-            let result = VerifiableCredential::create(issuer, credential_subject(), None);
-
-            match result {
-                Err(Web5Error::Parameter(err_msg)) => {
-                    assert_eq!(err_msg, "named issuer name must not be empty");
-                }
-                _ => panic!("Expected Web5Error::Parameter with specific error message"),
-            };
-        }
-
-        #[test]
-        fn test_issuer_object_must_be_set() {
-            TEST_SUITE.include(test_name!());
-
-            let issuer = Issuer::Object(ObjectIssuer {
-                id: ISSUER_DID_URI.to_string(),
-                name: "Example Name".to_string(),
-                additional_properties: None,
-            });
-
-            let vc =
-                VerifiableCredential::create(issuer.clone(), credential_subject(), None).unwrap();
-
-            assert_eq!(vc.issuer, issuer);
-        }
-
-        #[test]
-        fn test_issuer_object_supports_additional_properties() {
-            TEST_SUITE.include(test_name!());
-
-            let additional_properties = JsonObject {
-                properties: HashMap::from([(
-                    "extra_key".to_string(),
-                    JsonValue::String("extra_value".to_string()),
-                )]),
-            };
-
-            let issuer = Issuer::Object(ObjectIssuer {
-                id: ISSUER_DID_URI.to_string(),
-                name: "Example Name".to_string(),
-                additional_properties: Some(additional_properties.clone()),
-            });
-
-            let vc =
-                VerifiableCredential::create(issuer.clone(), credential_subject(), None).unwrap();
-
-            match vc.issuer {
-                Issuer::Object(ref obj) => {
-                    assert_eq!(obj.additional_properties, Some(additional_properties));
-                }
-                _ => panic!("Issuer is not an ObjectIssuer"),
-            };
-        }
-
-        #[test]
-        fn test_credential_subject_id_must_not_be_empty() {
-            TEST_SUITE.include(test_name!());
-
-            let credential_subject = CredentialSubject::from("");
-
-            let result = VerifiableCredential::create(issuer(), credential_subject, None);
-
-            match result {
-                Err(Web5Error::Parameter(err_msg)) => {
-                    assert_eq!(err_msg, "subject id must not be empty");
-                }
-                _ => panic!("Expected Web5Error::Parameter with specific error message"),
-            };
-        }
-
-        #[test]
-        fn test_credential_subject_must_be_set() {
-            TEST_SUITE.include(test_name!());
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), None).unwrap();
-
-            assert_eq!(vc.credential_subject, credential_subject());
-        }
-
-        #[test]
-        fn test_credential_subject_supports_additional_properties() {
-            TEST_SUITE.include(test_name!());
-
-            let additional_properties = JsonObject {
-                properties: HashMap::from([(
-                    "extra_key".to_string(),
-                    JsonValue::String("extra_value".to_string()),
-                )]),
-            };
-
-            let credential_subject = CredentialSubject {
-                id: SUBJECT_DID_URI.to_string(),
-                additional_properties: Some(additional_properties.clone()),
-            };
-
-            let vc =
-                VerifiableCredential::create(issuer(), credential_subject.clone(), None).unwrap();
-
-            assert_eq!(
-                vc.credential_subject.additional_properties,
-                Some(additional_properties)
-            );
-        }
-
-        #[test]
-        fn test_issuance_date_must_be_set() {
-            TEST_SUITE.include(test_name!());
-
-            let issuance_date = SystemTime::now();
-
-            let options = Some(VerifiableCredentialCreateOptions {
-                issuance_date: Some(issuance_date),
-                ..Default::default()
-            });
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), options).unwrap();
-
-            assert_eq!(vc.issuance_date, issuance_date);
-        }
-
-        #[test]
-        fn test_issuance_date_must_be_now_if_not_supplied() {
-            TEST_SUITE.include(test_name!());
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), None).unwrap();
-
-            let now = SystemTime::now();
-            let hundred_millis_ago = now - std::time::Duration::from_millis(100);
-
-            assert!(vc.issuance_date >= hundred_millis_ago && vc.issuance_date <= now);
-        }
-
-        #[test]
-        fn test_expiration_date_must_be_set_if_supplied() {
-            TEST_SUITE.include(test_name!());
-
-            let expiration_date = SystemTime::now();
-            let options = Some(VerifiableCredentialCreateOptions {
-                expiration_date: Some(expiration_date),
-                ..Default::default()
-            });
-
-            let vc = VerifiableCredential::create(issuer(), credential_subject(), options).unwrap();
-
-            assert_eq!(vc.expiration_date, Some(expiration_date));
-        }
-    }
 
     mod from_vc_jwt {
         use super::*;
+        use crate::{credentials::CredentialError, errors::Web5Error};
+        use crate::{
+            dids::resolution::resolution_metadata::ResolutionMetadataError,
+            test_helpers::UnitTestSuite, test_name,
+        };
+        use lazy_static::lazy_static;
 
         lazy_static! {
             static ref TEST_SUITE: UnitTestSuite =
                 UnitTestSuite::new("verifiable_credential_1_1_from_vc_jwt");
+        }
+
+        fn assert_credential_error<T>(result: Result<T>, expected_error: CredentialError)
+        where
+            T: std::fmt::Debug,
+        {
+            match result {
+                Err(Web5Error::CredentialError(err)) => {
+                    assert_eq!(err, expected_error, "Unexpected CredentialError variant");
+                }
+                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
+            };
         }
 
         #[test]
@@ -1011,12 +132,7 @@ mod tests {
 
             let result = VerifiableCredential::from_vc_jwt(vc_jwt_without_kid, true);
 
-            match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(err, CredentialError::MissingKid)
-                }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
-            };
+            assert_credential_error(result, CredentialError::MissingKid);
         }
 
         #[test]
@@ -1109,9 +225,9 @@ mod tests {
             let vc = VerifiableCredential::from_vc_jwt(vc_jwt_valid, true)
                 .expect("vc_jwt should be valid");
             assert_eq!(
-                "did:jwk:eyJhbGciOiJFZDI1NTE5Iiwia3R5IjoiT0tQIiwiY3J2IjoiRWQyNTUxOSIsIngiOiJHVzZFTDlITTltdHlycGlYdFFUMGpxZk52aWNnQTlBVDg0MHY1Y08yb1RrIn0#0", 
-                vc.issuer.to_string()
-            );
+              "did:jwk:eyJhbGciOiJFZDI1NTE5Iiwia3R5IjoiT0tQIiwiY3J2IjoiRWQyNTUxOSIsIngiOiJHVzZFTDlITTltdHlycGlYdFFUMGpxZk52aWNnQTlBVDg0MHY1Y08yb1RrIn0#0", 
+              vc.issuer.to_string()
+          );
         }
 
         #[test]
@@ -1148,9 +264,9 @@ mod tests {
             match vc.issuer {
                 Issuer::String(issuer) => {
                     assert_eq!(
-                        "did:jwk:eyJhbGciOiJFZDI1NTE5Iiwia3R5IjoiT0tQIiwiY3J2IjoiRWQyNTUxOSIsIngiOiJyd2hXSUNYclJ3b0ZaRmR1M0lsNi1BNGUtdjk3QlMxRkZRaVE4aWNmWktrIn0",
-                        issuer
-                    )
+                      "did:jwk:eyJhbGciOiJFZDI1NTE5Iiwia3R5IjoiT0tQIiwiY3J2IjoiRWQyNTUxOSIsIngiOiJyd2hXSUNYclJ3b0ZaRmR1M0lsNi1BNGUtdjk3QlMxRkZRaVE4aWNmWktrIn0",
+                      issuer
+                  )
                 }
                 Issuer::Object(_) => panic!("issuer should be string"),
             }
@@ -1168,9 +284,9 @@ mod tests {
                 Issuer::String(_) => panic!("issuer should be object"),
                 Issuer::Object(issuer) => {
                     assert_eq!(
-                        "did:jwk:eyJhbGciOiJFZDI1NTE5Iiwia3R5IjoiT0tQIiwiY3J2IjoiRWQyNTUxOSIsIngiOiI1Uk1yaUM1VlhubzZSVDhMWVVrbnpJZnNjaTQyYmxBaWlLWkpCZGhnVnVBIn0",
-                        issuer.id
-                    );
+                      "did:jwk:eyJhbGciOiJFZDI1NTE5Iiwia3R5IjoiT0tQIiwiY3J2IjoiRWQyNTUxOSIsIngiOiI1Uk1yaUM1VlhubzZSVDhMWVVrbnpJZnNjaTQyYmxBaWlLWkpCZGhnVnVBIn0",
+                      issuer.id
+                  );
                     assert_eq!("some name", issuer.name)
                 }
             }
@@ -1181,15 +297,8 @@ mod tests {
             TEST_SUITE.include(test_name!());
 
             let vc_jwt_with_missing_vc_claim = r#"eyJ0eXAiOiJKV1QiLCJhbGciOiJFZERTQSIsImtpZCI6ImRpZDpqd2s6ZXlKaGJHY2lPaUpGWkRJMU5URTVJaXdpYTNSNUlqb2lUMHRRSWl3aVkzSjJJam9pUldReU5UVXhPU0lzSW5naU9pSlNSbkZSVlVWS1RFOVhlbXh3T1ZaRk1rdEtSalp6UjBwT00yVnpaWHBsY0hSSE0ySTFlbTh4YjAwNEluMCMwIn0.eyJpc3MiOiJkaWQ6andrOmV5SmhiR2NpT2lKRlpESTFOVEU1SWl3aWEzUjVJam9pVDB0UUlpd2lZM0oySWpvaVJXUXlOVFV4T1NJc0luZ2lPaUpTUm5GUlZVVktURTlYZW14d09WWkZNa3RLUmpaelIwcE9NMlZ6WlhwbGNIUkhNMkkxZW04eGIwMDRJbjAiLCJqdGkiOiJ1cm46dXVpZDozNmU0ZjllNi0yYzdjLTQ0NGMtOTI4OS0zNDhmY2IxNDZlYjYiLCJzdWIiOiJkaWQ6ZGh0OnFnbW1weWp3NWh3bnFmZ3puN3dtcm0zM2FkeThnYjh6OWlkZWliNm05Z2o0eXM2d255OHkiLCJuYmYiOjE3MjQ4NTA1MjIsImlhdCI6MTcyNDg1MDUyMn0.SqwZC0q9RuHp9hAtFmE6sBYeJ1uHuuq1hyijF0NmW9nksSBqtDpfNroNlitK_Tl-CLWtwbTpK3b3JduTfzGEAw"#;
-
             let result = VerifiableCredential::from_vc_jwt(vc_jwt_with_missing_vc_claim, true);
-
-            match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(err, CredentialError::MissingClaim("vc".to_string()))
-                }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
-            };
+            assert_credential_error(result, CredentialError::MissingClaim("vc".to_string()));
         }
 
         #[test]
@@ -1197,15 +306,8 @@ mod tests {
             TEST_SUITE.include(test_name!());
 
             let vc_jwt_with_missing_jti_claim = r#"eyJ0eXAiOiJKV1QiLCJhbGciOiJFZERTQSIsImtpZCI6ImRpZDpqd2s6ZXlKaGJHY2lPaUpGWkRJMU5URTVJaXdpYTNSNUlqb2lUMHRRSWl3aVkzSjJJam9pUldReU5UVXhPU0lzSW5naU9pSm5jMjlTZGsxUFlXMHliMlJQTlY4NWVqbExlV2xzV1VzM1Yzb3RZa1owWW5wdlVrWm1iVTlUTVRJNEluMCMwIn0.eyJ2YyI6eyJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSJdLCJpZCI6InVybjp1dWlkOjEwODM2MzgwLWI2MmMtNGVmZC04YmU0LTZhNzJiMDZjYWI4NyIsInR5cGUiOlsiVmVyaWZpYWJsZUNyZWRlbnRpYWwiXSwiaXNzdWVyIjoiZGlkOmp3azpleUpoYkdjaU9pSkZaREkxTlRFNUlpd2lhM1I1SWpvaVQwdFFJaXdpWTNKMklqb2lSV1F5TlRVeE9TSXNJbmdpT2lKbmMyOVNkazFQWVcweWIyUlBOVjg1ZWpsTGVXbHNXVXMzVjNvdFlrWjBZbnB2VWtabWJVOVRNVEk0SW4wIiwiaXNzdWFuY2VEYXRlIjoiMjAyNC0wOC0yOFQxMzoxMDo1NS4yMDYwOTIrMDA6MDAiLCJleHBpcmF0aW9uRGF0ZSI6bnVsbCwiY3JlZGVudGlhbFN1YmplY3QiOnsiaWQiOiJkaWQ6ZGh0OnFnbW1weWp3NWh3bnFmZ3puN3dtcm0zM2FkeThnYjh6OWlkZWliNm05Z2o0eXM2d255OHkifX0sImlzcyI6ImRpZDpqd2s6ZXlKaGJHY2lPaUpGWkRJMU5URTVJaXdpYTNSNUlqb2lUMHRRSWl3aVkzSjJJam9pUldReU5UVXhPU0lzSW5naU9pSm5jMjlTZGsxUFlXMHliMlJQTlY4NWVqbExlV2xzV1VzM1Yzb3RZa1owWW5wdlVrWm1iVTlUTVRJNEluMCIsInN1YiI6ImRpZDpkaHQ6cWdtbXB5anc1aHducWZnem43d21ybTMzYWR5OGdiOHo5aWRlaWI2bTlnajR5czZ3bnk4eSIsIm5iZiI6MTcyNDg1MDY1NSwiaWF0IjoxNzI0ODUwNjU1fQ.1XDmdvB1GDsCHw9Qwp0HA5r8W-JnZB4lz9Yqo0C2V_EEe-uk88bQSl8P9HV8ViNyBC_YaYatLiPTD4jBZY77DA"#;
-
             let result = VerifiableCredential::from_vc_jwt(vc_jwt_with_missing_jti_claim, true);
-
-            match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(err, CredentialError::MissingClaim("jti".to_string()))
-                }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
-            };
+            assert_credential_error(result, CredentialError::MissingClaim("jti".to_string()));
         }
 
         #[test]
@@ -1344,6 +446,8 @@ mod tests {
             };
         }
 
+        // --- Data Model Validation Tests ---
+
         #[test]
         fn test_validate_dm_empty_id() {
             TEST_SUITE.include(test_name!());
@@ -1353,13 +457,13 @@ mod tests {
             let result = VerifiableCredential::from_vc_jwt(vc_jwt_with_empty_id, true);
 
             match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(
-                        err,
-                        CredentialError::DataModelValidationError("missing id".to_string())
-                    )
+                Err(Web5Error::CredentialError(CredentialError::DataModelValidationError(msg))) => {
+                    assert_eq!(msg, "missing id")
                 }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
+                _ => panic!(
+                    "Expected CredentialError::DataModelValidationError, but got: {:?}",
+                    result
+                ),
             };
         }
 
@@ -1372,13 +476,13 @@ mod tests {
             let result = VerifiableCredential::from_vc_jwt(vc_jwt_with_empty_context, true);
 
             match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(
-                        err,
-                        CredentialError::DataModelValidationError("missing context".to_string())
-                    )
+                Err(Web5Error::CredentialError(CredentialError::DataModelValidationError(msg))) => {
+                    assert_eq!(msg, "missing context")
                 }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
+                _ => panic!(
+                    "Expected CredentialError::DataModelValidationError, but got: {:?}",
+                    result
+                ),
             };
         }
 
@@ -1387,17 +491,16 @@ mod tests {
             TEST_SUITE.include(test_name!());
 
             let vc_jwt_without_base_context = r#"eyJ0eXAiOiJKV1QiLCJhbGciOiJFZERTQSIsImtpZCI6ImRpZDpqd2s6ZXlKaGJHY2lPaUpGWkRJMU5URTVJaXdpYTNSNUlqb2lUMHRRSWl3aVkzSjJJam9pUldReU5UVXhPU0lzSW5naU9pSnFSVWRmU1V4TlZDMVVUMGgyTjIxeFJ6UlJWVmRMTWs1dU9FcGlUVU5OWldKQ1pXVjRkVGxIYlhWWkluMCMwIn0.eyJ2YyI6eyJAY29udGV4dCI6WyJhIGNvbnRleHQiXSwiaWQiOiJ1cm46dXVpZDo4N2VmMDI1MC0yYWE2LTQyNTctYjIxMi0xYzAyMWFhZDY2Y2YiLCJ0eXBlIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsIl0sImlzc3VlciI6ImRpZDpqd2s6ZXlKaGJHY2lPaUpGWkRJMU5URTVJaXdpYTNSNUlqb2lUMHRRSWl3aVkzSjJJam9pUldReU5UVXhPU0lzSW5naU9pSnFSVWRmU1V4TlZDMVVUMGgyTjIxeFJ6UlJWVmRMTWs1dU9FcGlUVU5OWldKQ1pXVjRkVGxIYlhWWkluMCIsImlzc3VhbmNlRGF0ZSI6IjIwMjQtMDgtMjhUMTM6MzQ6MjcuODk4MDkwKzAwOjAwIiwiZXhwaXJhdGlvbkRhdGUiOm51bGwsImNyZWRlbnRpYWxTdWJqZWN0Ijp7ImlkIjoiZGlkOmRodDpxZ21tcHlqdzVod25xZmd6bjd3bXJtMzNhZHk4Z2I4ejlpZGVpYjZtOWdqNHlzNndueTh5In19LCJpc3MiOiJkaWQ6andrOmV5SmhiR2NpT2lKRlpESTFOVEU1SWl3aWEzUjVJam9pVDB0UUlpd2lZM0oySWpvaVJXUXlOVFV4T1NJc0luZ2lPaUpxUlVkZlNVeE5WQzFVVDBoMk4yMXhSelJSVlZkTE1rNXVPRXBpVFVOTlpXSkNaV1Y0ZFRsSGJYVlpJbjAiLCJqdGkiOiJ1cm46dXVpZDo4N2VmMDI1MC0yYWE2LTQyNTctYjIxMi0xYzAyMWFhZDY2Y2YiLCJzdWIiOiJkaWQ6ZGh0OnFnbW1weWp3NWh3bnFmZ3puN3dtcm0zM2FkeThnYjh6OWlkZWliNm05Z2o0eXM2d255OHkiLCJuYmYiOjE3MjQ4NTIwNjcsImlhdCI6MTcyNDg1MjA2N30.cgkGQF5CXqHw_C1KNaKLFeIzPzmBuWRzRk-7KgvEYc1jJzwoXoOWB6cn-8I3MjWAgd_NeM1Yt656lJ60gy0RAQ"#;
-
             let result = VerifiableCredential::from_vc_jwt(vc_jwt_without_base_context, true);
 
             match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(
-                        err,
-                        CredentialError::DataModelValidationError("missing context".to_string())
-                    )
+                Err(Web5Error::CredentialError(CredentialError::DataModelValidationError(msg))) => {
+                    assert_eq!(msg, "missing context")
                 }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
+                _ => panic!(
+                    "Expected CredentialError::DataModelValidationError, but got: {:?}",
+                    result
+                ),
             };
         }
 
@@ -1410,13 +513,13 @@ mod tests {
             let result = VerifiableCredential::from_vc_jwt(vc_jwt_with_empty_type, true);
 
             match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(
-                        err,
-                        CredentialError::DataModelValidationError("missing type".to_string())
-                    )
+                Err(Web5Error::CredentialError(CredentialError::DataModelValidationError(msg))) => {
+                    assert_eq!(msg, "missing type")
                 }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
+                _ => panic!(
+                    "Expected CredentialError::DataModelValidationError, but got: {:?}",
+                    result
+                ),
             };
         }
 
@@ -1429,13 +532,13 @@ mod tests {
             let result = VerifiableCredential::from_vc_jwt(vc_jwt_without_base_type, true);
 
             match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(
-                        err,
-                        CredentialError::DataModelValidationError("missing type".to_string())
-                    )
+                Err(Web5Error::CredentialError(CredentialError::DataModelValidationError(msg))) => {
+                    assert_eq!(msg, "missing type")
                 }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
+                _ => panic!(
+                    "Expected CredentialError::DataModelValidationError, but got: {:?}",
+                    result
+                ),
             };
         }
 
@@ -1445,16 +548,17 @@ mod tests {
 
             let vc_jwt_with_empty_issuer = r#"eyJ0eXAiOiJKV1QiLCJhbGciOiJFZERTQSIsImtpZCI6ImRpZDpqd2s6ZXlKaGJHY2lPaUpGWkRJMU5URTVJaXdpYTNSNUlqb2lUMHRRSWl3aVkzSjJJam9pUldReU5UVXhPU0lzSW5naU9pSnRPSEZRTVhoR1p6RndZMlpqZUY5UWVrUnJOMjFPYVhoak9YQTFWamN4WlVZelYwRTViSGwzTWpsckluMCMwIn0.eyJ2YyI6eyJAY29udGV4dCI6WyJodHRwczovL3d3dy53My5vcmcvMjAxOC9jcmVkZW50aWFscy92MSJdLCJpZCI6InVybjp1dWlkOmFhMmNjNWNkLTg4N2QtNDFkMi1iZTM3LTIzMjMxMGVkODdjMiIsInR5cGUiOlsiVmVyaWZpYWJsZUNyZWRlbnRpYWwiXSwiaXNzdWVyIjoiIiwiaXNzdWFuY2VEYXRlIjoiMjAyNC0wOC0yOFQxMzozOToxNS42MjMzOTMrMDA6MDAiLCJleHBpcmF0aW9uRGF0ZSI6bnVsbCwiY3JlZGVudGlhbFN1YmplY3QiOnsiaWQiOiJkaWQ6ZGh0OnFnbW1weWp3NWh3bnFmZ3puN3dtcm0zM2FkeThnYjh6OWlkZWliNm05Z2o0eXM2d255OHkifX0sImlzcyI6IiIsImp0aSI6InVybjp1dWlkOmFhMmNjNWNkLTg4N2QtNDFkMi1iZTM3LTIzMjMxMGVkODdjMiIsInN1YiI6ImRpZDpkaHQ6cWdtbXB5anc1aHducWZnem43d21ybTMzYWR5OGdiOHo5aWRlaWI2bTlnajR5czZ3bnk4eSIsIm5iZiI6MTcyNDg1MjM1NSwiaWF0IjoxNzI0ODUyMzU1fQ.mRYZKF3qNz_Vyg8xpemuBOipGLOliYy9xJ6b9ZqcMNjZbb8GtEyiaBv8rgF2jqmHreRT71wHaT3P6mV9GsQOCA"#;
 
-            let result = VerifiableCredential::from_vc_jwt(vc_jwt_with_empty_issuer, true);
+            let vc = decode(vc_jwt_with_empty_issuer, true).unwrap();
+            let result = validate_vc_data_model(&vc);
 
             match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(
-                        err,
-                        CredentialError::DataModelValidationError("missing issuer".to_string())
-                    )
+                Err(CredentialError::DataModelValidationError(msg)) => {
+                    assert_eq!(msg, "missing issuer")
                 }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
+                _ => panic!(
+                    "Expected CredentialError::DataModelValidationError, but got: {:?}",
+                    result
+                ),
             };
         }
 
@@ -1467,15 +571,13 @@ mod tests {
             let result = VerifiableCredential::from_vc_jwt(vc_jwt_with_empty_subject, true);
 
             match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(
-                        err,
-                        CredentialError::DataModelValidationError(
-                            "missing credential subject".to_string()
-                        )
-                    )
+                Err(Web5Error::CredentialError(CredentialError::DataModelValidationError(msg))) => {
+                    assert_eq!(msg, "missing credential subject")
                 }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
+                _ => panic!(
+                    "Expected CredentialError::DataModelValidationError, but got: {:?}",
+                    result
+                ),
             };
         }
 
@@ -1489,15 +591,13 @@ mod tests {
                 VerifiableCredential::from_vc_jwt(vc_jwt_with_issuance_date_in_future, true);
 
             match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(
-                        err,
-                        CredentialError::DataModelValidationError(
-                            "issuance date in future".to_string()
-                        )
-                    )
+                Err(Web5Error::CredentialError(CredentialError::DataModelValidationError(msg))) => {
+                    assert_eq!(msg, "issuance date in future")
                 }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
+                _ => panic!(
+                    "Expected CredentialError::DataModelValidationError, but got: {:?}",
+                    result
+                ),
             };
         }
 
@@ -1510,148 +610,14 @@ mod tests {
             let result = VerifiableCredential::from_vc_jwt(vc_jwt_with_expired, true);
 
             match result {
-                Err(Web5Error::CredentialError(err)) => {
-                    assert_eq!(
-                        err,
-                        CredentialError::DataModelValidationError("credential expired".to_string())
-                    )
+                Err(Web5Error::CredentialError(CredentialError::DataModelValidationError(msg))) => {
+                    assert_eq!(msg, "credential expired")
                 }
-                _ => panic!("Expected Web5Error::CredentialError, but got: {:?}", result),
+                _ => panic!(
+                    "Expected CredentialError::DataModelValidationError, but got: {:?}",
+                    result
+                ),
             };
-        }
-    }
-
-    mod sign {
-        use crate::dids::methods::did_jwk::DidJwk;
-
-        use super::*;
-
-        lazy_static! {
-            static ref TEST_SUITE: UnitTestSuite =
-                UnitTestSuite::new("verifiable_credential_1_1_sign");
-        }
-
-        #[test]
-        fn z_assert_all_suite_cases_covered() {
-            // fn name prefixed with `z_*` b/c rust test harness executes in alphabetical order,
-            // unless intentionally executed with "shuffle" https://doc.rust-lang.org/rustc/tests/index.html#--shuffle
-            // this may not work if shuffled or if test list grows to the extent of 100ms being insufficient wait time
-
-            // wait 100ms to be last-in-queue of mutex lock
-            std::thread::sleep(std::time::Duration::from_millis(100));
-
-            TEST_SUITE.assert_coverage()
-        }
-
-        #[test]
-        fn test_can_sign_then_verify() {
-            TEST_SUITE.include(test_name!());
-
-            let bearer_did = DidJwk::create(None).unwrap();
-            let vc = VerifiableCredential::create(
-                Issuer::String(bearer_did.did.uri.clone()),
-                credential_subject(),
-                None,
-            )
-            .unwrap();
-
-            let vc_jwt = vc
-                .sign(&bearer_did, None)
-                .expect("should be able to sign vc jwt");
-
-            let vc_from_vc_jwt = VerifiableCredential::from_vc_jwt(&vc_jwt, true)
-                .expect("should be able to verify the signed vc jwt");
-            assert_eq!(vc.id, vc_from_vc_jwt.id)
-        }
-
-        #[test]
-        fn test_bearer_did_mismatch_issuer() {
-            TEST_SUITE.include(test_name!());
-
-            let bearer_did = DidJwk::create(None).unwrap();
-            let vc = VerifiableCredential::create(
-                Issuer::String(bearer_did.did.uri.clone()),
-                credential_subject(),
-                None,
-            )
-            .unwrap();
-
-            let different_bearer_did = DidJwk::create(None).unwrap();
-            let result = vc.sign(&different_bearer_did, None);
-
-            match result {
-                Err(Web5Error::Parameter(err_msg)) => {
-                    assert_eq!(
-                        err_msg,
-                        format!(
-                            "bearer_did uri {} does not match issuer {}",
-                            different_bearer_did.did.uri, bearer_did.did.uri
-                        )
-                    )
-                }
-                _ => panic!("Expected Web5Error::Parameter, but got: {:?}", result),
-            };
-        }
-
-        #[test]
-        fn test_defaults_to_first_vm() {
-            TEST_SUITE.include(test_name!());
-
-            let bearer_did = DidJwk::create(None).unwrap();
-            let vc = VerifiableCredential::create(
-                Issuer::String(bearer_did.did.uri.clone()),
-                credential_subject(),
-                None,
-            )
-            .unwrap();
-
-            let vc_jwt = vc
-                .sign(&bearer_did, None)
-                .expect("should sign with default vm");
-
-            let header = josekit::jwt::decode_header(vc_jwt).unwrap();
-
-            let kid = header
-                .claim("kid")
-                .and_then(serde_json::Value::as_str)
-                .unwrap();
-
-            assert_eq!(bearer_did.document.verification_method[0].id, kid)
-        }
-
-        #[test]
-        fn test_vm_must_be_assertion_method() {
-            TEST_SUITE.include(test_name!());
-
-            let mut bearer_did = DidJwk::create(None).unwrap();
-            let vc = VerifiableCredential::create(
-                Issuer::String(bearer_did.did.uri.clone()),
-                credential_subject(),
-                None,
-            )
-            .unwrap();
-
-            // remove the assertionMethod
-            if let Some(assertion_method) = bearer_did.document.assertion_method.as_mut() {
-                assertion_method.remove(0);
-            }
-
-            let vm_id = bearer_did.document.verification_method[0].id.clone();
-
-            let result = vc.sign(&bearer_did, Some(vm_id.clone()));
-
-            match result {
-                Err(Web5Error::Parameter(err_msg)) => {
-                    assert_eq!(
-                        err_msg,
-                        format!(
-                            "verification_method_id {} is not an assertion_method",
-                            vm_id
-                        )
-                    )
-                }
-                _ => panic!("expected Web5Error::Parameter but got {:?}", result),
-            }
         }
     }
 }
